Provide a RouteEnhancerCollectorTrait for easier use of route enhancers.
