+++ composer.json
@@ -13,7 +13,7 @@
     ],
     "minimum-stability": "dev",
     "require": {
-        "php": ">=5.3.9",
+        "php": ">=5.4.0",
         "symfony/routing": "^2.2|3.*",
         "symfony/http-kernel": "^2.2|3.*",
         "psr/log": "1.*"
